% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_survey.R
\name{add_survey}
\alias{add_survey}
\title{Add a survey}
\usage{
add_survey(date, observers, lead_observer, space, space_species, root)
}
\arguments{
\item{date}{the date of the survey}

\item{observers}{a vector of observer ids.}

\item{lead_observer}{the observer id of the lead observer.}

\item{space}{a vector of space codes.}

\item{space_species}{a data.frame with space, species and count}

\item{root}{Either a \code{git_repository} object or the path to a git repository}
}
\description{
Add a survey
}
