% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_observations.R
\name{aggregate_observations}
\alias{aggregate_observations}
\title{Aggregate observations at a species level}
\usage{
aggregate_observations(species_code, object_name, root)
}
\arguments{
\item{object_name}{the name of the object}

\item{root}{Either a \code{git_repository} object or the path to a git repository}

\item{species}{id of the species}
}
\description{
Aggregate observations at a species level
}
