% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_object.R
\name{display_object}
\alias{display_object}
\title{Display the object}
\usage{
display_object(
  object_name,
  root,
  extra_data,
  extra_var,
  legend_title = extra_var,
  labels = waiver(),
  breaks = waiver(),
  shift = TRUE,
  rotate = TRUE,
  crs = 31370
)
}
\arguments{
\item{root}{Either a \code{git_repository} object or the path to a git repository}

\item{extra_data}{an optional data.frame to fill the spaces.}

\item{extra_var}{the name of the variable in \code{extra_data} to fill the spaces.}

\item{legend_title}{title for the colour legend}

\item{shift}{shift parts with different floors to avoid overlap.
Defaults to \code{TRUE}.}

\item{rotate}{Rotate the part along the general direction of the object.
Default to \code{TRUE}.}

\item{crs}{target coordinate reference system: object of class `crs`, or input string for \link[sf]{st_crs}}
}
\description{
Display the object
}
