% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_part.R
\name{display_part}
\alias{display_part}
\title{Display the floor plan of a part}
\usage{
display_part(
  part_id,
  root,
  label = c("code", "id", "none"),
  connection = FALSE,
  rotate = TRUE,
  crs = 31370
)
}
\arguments{
\item{part_id}{the id of the part}

\item{root}{Either a \code{git_repository} object or the path to a git repository}

\item{label}{Which labels to display on the labelled spaces?
\code{code} is the short code, \code{ìd} is the internal id, \code{none} hides the labels.}

\item{connection}{Display optional connections between the centroids.}

\item{rotate}{Rotate the part along the general direction of the object.
Default to \code{TRUE}.}

\item{crs}{target coordinate reference system: object of class `crs`, or input string for \link[sf]{st_crs}}
}
\description{
Display the floor plan of a part
}
