% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_surveyed.R
\name{display_surveyed}
\alias{display_surveyed}
\title{Display a binary map with the surveyed spaces}
\usage{
display_surveyed(
  object_name,
  year,
  root,
  legend_title,
  labels = waiver(),
  breaks = waiver(),
  crs = 31370
)
}
\arguments{
\item{object_name}{the name of the object}

\item{year}{Year of the survey}

\item{root}{Either a \code{git_repository} object or the path to a git repository}

\item{legend_title}{title for the colour legend}

\item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{crs}{target coordinate reference system: object of class `crs`, or input string for \link[sf]{st_crs}}
}
\description{
Display a binary map with the surveyed spaces
}
