% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_network}
\alias{get_network}
\title{Create a network for the connected spaces}
\usage{
get_network(object_name, direct_only = FALSE, root)
}
\arguments{
\item{object_name}{the name of the object}

\item{root}{Either a \code{git_repository} object or the path to a git repository}
}
\description{
Create a network for the connected spaces
}
