% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_model.R
\name{imputation_model}
\alias{imputation_model}
\title{Fit imputation model}
\usage{
imputation_model(species_code, object_name, root, direct_only = FALSE)
}
\arguments{
\item{species_code}{code of the species}

\item{object_name}{the name of the object}

\item{root}{Either a \code{git_repository} object or the path to a git repository}

\item{direct_only}{When \code{TRUE}, use only direct connections (through existing
openings).
When \code{FALSE} use also connections to the nearest space through a wall.
Defaults to \code{FALSE}}
}
\description{
Fit imputation model
}
