% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{impute missing observations}
\usage{
impute(model, level = c("object", "part", "space"), n_sim = 999)
}
\arguments{
\item{model}{the output of \code{ìmputation_model()}.}

\item{level}{the required level of aggregation.
Defaults to \code{"object"}.}

\item{n_sim}{number of simulations.}
}
\description{
impute missing observations
}
