% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_display.R
\name{scale_display}
\alias{scale_display}
\title{Scale a display
Return a vector of the optimal width and height}
\usage{
scale_display(display, scale = 200, width = 180, height = 260)
}
\arguments{
\item{display}{the output of \code{display_object()}, \code{display_part()} or
\code{display_importance()}.}

\item{scale}{The minimal scale.
This prevents small objects to be scaled up.
Decreasing the number will enlarge small objects.
Defaults to \code{200}, referring to a 1:200 scale.}

\item{width}{maximal width in mm.}

\item{height}{maximal height in mm.}
}
\description{
Scale a display
Return a vector of the optimal width and height
}
