% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_object.R
\name{display_object}
\alias{display_object}
\title{Display the object}
\usage{
display_object(
  object_name,
  root,
  extra_data,
  extra_var,
  legend_title = extra_var,
  labels = waiver(),
  breaks = waiver(),
  shift = TRUE,
  rotate = TRUE,
  crs = 31370
)
}
\arguments{
\item{object_name}{the name of the object}

\item{root}{Either a \code{git_repository} object or the path to a git repository}

\item{extra_data}{an optional data.frame to fill the spaces.}

\item{extra_var}{the name of the variable in \code{extra_data} to fill the spaces.}

\item{legend_title}{title for the colour legend}

\item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks (the scale limits)
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{shift}{shift parts with different floors to avoid overlap.
Defaults to \code{TRUE}.}

\item{rotate}{Rotate the part along the general direction of the object.
Default to \code{TRUE}.}

\item{crs}{target coordinate reference system: object of class `crs`, or input string for \link[sf]{st_crs}}
}
\description{
Display the object
}
