% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validata the raw data repository}
\usage{
validate_data(root)
}
\arguments{
\item{root}{Either a \code{git_repository} object or the path to a git repository}
}
\description{
Validata the raw data repository
}
